add dry run mode to get set-of-mark plots

+++ compete_and_select/evaluate_object_selection.py
@@ -8,6 +8,7 @@ from .sam import boxes_to_masks
 from .clip_feature_extraction import get_text_embeds, get_clip_embeddings, get_full_scale_clip_embeddings
 from .lmp_scene_api import get_selection_policy
 
+import matplotlib.pyplot as plt
 import json
 import PIL.Image
 import pickle
@@ -16,19 +17,13 @@ import torch
 from torch.nn.functional import interpolate
 import numpy as np
 from .standalone_compete_and_select import select_with_vlm, describe_objects
+from .class_labels import class_labels
 
-slug_labels = ['blue_mug', 'white_mug', 'green_coffee_cup', 'white_plastic_cup', 'uva_football_cup', 'green_cup']
+gp = 'condiments'
+natural_language_labels = class_labels[gp]['natural_language_labels']
+slug_labels = class_labels[gp]['slug_labels']
 
-natural_language_labels = [
-    "blue mug",
-    "white mug",
-    "green coffee cup",
-    "white plastic cup",
-    "UVA football cup",
-    "green plastic cup",
-]
-
-folder = './photos/solidbg/cups'
+folder = './photos/solidbg/' + gp
 
 with open(os.path.join(folder, "expected.json")) as f:
     expected_per_img = json.load(f)
@@ -57,7 +52,7 @@ def generate_clip_embeddings():
     with open(os.path.join(detector_results_folder, "clip_embeddings.pkl"), "wb") as f:
         pickle.dump(clip_embeddings, f)
 
-def generate_direct_owlv2():
+def generate_direct_owlv2(coarse_label=None):
     # store in raw python format
     detections = {}
 
@@ -75,7 +70,8 @@ def generate_direct_owlv2():
 
         # go through natural language labels
         for k, label in enumerate(natural_language_labels):
-            dets = detect(image, label, use_clip_projection=True)
+            dets = detect(image, coarse_label or label, use_clip_projection=True)
+            # dets = detect(image, label, use_clip_projection=True)
 
             print(f"Detecting [{label}]: # = {len(dets)}")
 
@@ -105,52 +101,51 @@ def mask_iou(m1, m2):
     return (m1 & m2).sum() / (m1 | m2).sum()
 
 # cache the results to preserve credits.
-def generate_vlm_results__(image, detections_for_file, allow_descriptions=True):
-    vlm_outputs = []
-    
-    for detection_group in detections_for_file:
-        print("Running on detection group:", detection_group['label'])
-        bboxes = [det['bbox'] for det in detection_group['detections']]
-        
-        if allow_descriptions:
-            descriptions = describe_objects(image, bounding_boxes)
-        else:
-            descriptions = None
-        
-        # has "reasoning", "logits", and "response" keys
-        # returns a list of logits, corresponding to bboxes.
-        vlm_output = select_with_vlm(image, bboxes, detection_group['natural_language_label'], descriptions)
-        vlm_outputs.append(vlm_output)
-        
-    return vlm_outputs
-
-def generate_vlm_results(image, detections_for_file, allow_descriptions=True):
+def generate_vlm_results(image, detections_for_file, allow_descriptions, used_coarse_detections, dry_run):
     from concurrent.futures import ThreadPoolExecutor
+    
+    if dry_run:
+        print("==> Dry Run <== (Only for debugging, not making LM calls)")
+        describe_objects = lambda *_: None
+    
+    if used_coarse_detections and allow_descriptions and not dry_run:
+        bboxes = [d['bbox'] for d in detections_for_file[0]['detections']]
+        coarse_descriptions = describe_objects(image, bboxes)
+    else:
+        coarse_descriptions = None
     
     def inner(detection_group):
         print("Running on detection group:", detection_group['label'])
         bboxes = [det['bbox'] for det in detection_group['detections']]
         
         if allow_descriptions:
-            descriptions = describe_objects(image, bounding_boxes)
+            if coarse_descriptions:
+                # reuse coarse descriptions
+                descriptions = coarse_descriptions
+            elif not dry_run:
+                descriptions = describe_objects(image, bboxes)
         else:
             descriptions = None
+            
+        print("Allow Descriptions:", descriptions is not None)
         
         # has "reasoning", "logits", and "response" keys
         # returns a list of logits, corresponding to bboxes.
-        return select_with_vlm(image, bboxes, detection_group['natural_language_label'], descriptions)
+        return select_with_vlm(image, bboxes, detection_group['natural_language_label'], descriptions, dry_run=dry_run)
     
     with ThreadPoolExecutor() as ex:
         return list(ex.map(inner, detections_for_file))
 
-def obtain_vlm_results(allow_descriptions):
+def obtain_vlm_results(allow_descriptions, used_coarse_detections, dry_run=False):
     # may use the cache. REQUIRES: owlv2_direct.pkl
     detector_results_folder = obtain_detector_results_folder(folder)
     out_path = os.path.join(detector_results_folder, f"owlv2_direct__vlm_outputs__{'yes' if allow_descriptions else 'no'}_descriptions.pkl")
     
-    if os.path.exists(out_path):
-        with open(out_path, "rb") as f:
-            return pickle.load(f)
+    if not dry_run:
+        if os.path.exists(out_path):
+            with open(out_path, "rb") as f:
+                print("[cache hit: VLM results]")
+                return pickle.load(f)
         
     with open(os.path.join(detector_results_folder, "owlv2_direct.pkl"), "rb") as f:
         detections = pickle.load(f)
@@ -162,10 +157,11 @@ def obtain_vlm_results(allow_descriptions):
         full_filename = os.path.join(folder, filename)
         image = PIL.Image.open(full_filename).convert("RGB")
         detections_for_file = detections[filename]
-        generations[filename] = generate_vlm_results(image, detections_for_file, allow_descriptions)
+        generations[filename] = generate_vlm_results(image, detections_for_file, allow_descriptions, used_coarse_detections, dry_run)
         
-    with open(out_path, "wb") as f:
-        pickle.dump(generations, f)
+    if not dry_run:
+        with open(out_path, "wb") as f:
+            pickle.dump(generations, f)
         
     return generations
         
@@ -178,6 +174,8 @@ def evaluate_detection_results():
         
     with open(os.path.join(folder, "detector_results/owlv2_direct__vlm_outputs__yes_descriptions.pkl"), "rb") as f:
         vlm_results_yes_descriptions = pickle.load(f)
+    with open(os.path.join(folder, "detector_results/owlv2_direct__vlm_outputs__no_descriptions.pkl"), "rb") as f:
+        vlm_results_no_descriptions = pickle.load(f)
       
     # This is a lot of I/O for not a lot of compute
     # with open(os.path.join(folder, "detector_results/clip_embeddings.pkl"), "rb") as f:
@@ -193,8 +191,13 @@ def evaluate_detection_results():
         'iou_clip_cropped_score': [],
     }
     
-    text_embeds = get_text_embeds(["a photo of a " + label for label in natural_language_labels])
+    text_embeds = get_text_embeds(["a photo of " + label for label in natural_language_labels])
     text_embeds = {slug_label: embed for (slug_label, embed) in zip(slug_labels, text_embeds)}
+    
+    print(text_embeds.keys())
+    
+    counter = 0
+    # os.mkdir('plots', exist_ok=True)
     
     for filename in sorted(detections.keys()):
         print("Opening", filename)
@@ -207,6 +210,7 @@ def evaluate_detection_results():
         detections_for_file = detections[filename]
         
         vlm_results_yes_descriptions_for_file = vlm_results_yes_descriptions[filename]
+        vlm_results_no_descriptions_for_file = vlm_results_no_descriptions[filename]
         
         for expected_label, target_mask in zip(expected_for_file, masks[filename]):
             text_embed = text_embeds[expected_label]
@@ -217,9 +221,22 @@ def evaluate_detection_results():
             index_in_detections = [i for i in range(len(detections_for_file)) if detections_for_file[i]['label'] == expected_label][0]
             dets = detections_for_file[index_in_detections]['detections']
             
-            vlm_pred_index = np.argmax(vlm_results_yes_descriptions_for_file[index_in_detections]['logits'])
-            best_det_vlm_pred = dets[vlm_pred_index]
-            iou_vlm = mask_iou(best_det_vlm_pred['mask'], target_mask)
+            vlm_yes_pred_index = np.argmax(vlm_results_yes_descriptions_for_file[index_in_detections]['logits'])
+            best_det_vlm_yes_pred = dets[vlm_yes_pred_index]
+            iou_vlm_yes = mask_iou(best_det_vlm_yes_pred['mask'], target_mask)
+            
+            vlm_no_pred_index = np.argmax(vlm_results_no_descriptions_for_file[index_in_detections]['logits'])
+            best_det_vlm_no_pred = dets[vlm_no_pred_index]
+            iou_vlm_no = mask_iou(best_det_vlm_no_pred['mask'], target_mask)
+            
+            # draw the vlm pred mask
+            counter += 1
+            # plt.clf()
+            # plt.title("Result: " + filename + " => " + expected_label)
+            # plt.imshow(image)
+            # plt.imshow(target_mask.astype(float), alpha=target_mask.astype(float))
+            # # plt.imshow(best_det_vlm_pred['mask'].astype(float), alpha=best_det_vlm_pred['mask'].astype(float))
+            # plt.savefig(f'plots/plot_{counter}.png')
             
             # take the score argmax
             best_det_owlv2_score = max(dets, key=lambda det: det['score'])
@@ -232,9 +249,8 @@ def evaluate_detection_results():
             best_det_clip_mapped_score = max(dets, key=lambda det: clip_embeds[det['mask']].mean(axis=0) @ text_embed)
             iou_clip_mapped = mask_iou(best_det_clip_mapped_score['mask'], target_mask)
             
-            print(f"IOU[vlm]: {iou_vlm:.2f}, IOU [owlv2]: {iou_owlv2:.2f}, IOU [clip_cropped]: {iou_clip_cropped:.2f}, IOU [clip_mapped]: {iou_clip_mapped:.2f}")
+            print(f"IOU[c+s]: {iou_vlm_yes:.2f}, IOU[som]: {iou_vlm_no:.2f}, IOU [owlv2]: {iou_owlv2:.2f}, IOU [clip_cropped]: {iou_clip_cropped:.2f}, IOU [clip_mapped]: {iou_clip_mapped:.2f}")
         
-# generate_direct_owlv2()
-# generate_clip_embeddings()
-# obtain_vlm_results(allow_descriptions=True)
+# generate_direct_owlv2(coarse_label='spice container')
+obtain_vlm_results(allow_descriptions=False, used_coarse_detections=True, dry_run=True)
 evaluate_detection_results()
